Data for new cases of COVID19 from 22/01/2020 to 06/10/2020

A dataset contains the number of COVID19 cases worldwide from 22/01/2020 to 06/10/2020. Data is in longer format, with variable vars contains 6 types of cases: Confirmed, deceased, recovered, new daily confirmed, new daily death and new daily recovered.

data_total_newdaily

Format

A tibble with 292152 observations and 6 variables

Country/Region: Country/ Region name

date: The date of observation, using Date class.

lat: The latitude code.

long: The longitude code.

var: The cases type, c("Confirmed cases","Recovered cases","Deceased cases", "New daily cases","New death cases","New recovered cases")

value - The number of cases

Source

John Hopkins Github